+++ frontend/.prettierignore
@@ -2,3 +2,4 @@
 package.json
 package-lock.json
 public
+venv
