+++ frontend/.stylelintignore
@@ -1,0 +1,1 @@
+venv/**/*.css
