+++ frontend/gatsby-ssr.js
@@ -4,9 +4,9 @@
  * See: https://www.gatsbyjs.org/docs/ssr-apis/
  */
 
-import React from "react"
-import { Auth0Provider } from "./src/contexts/auth0Context"
+import React from "react";
+import { Auth0Provider } from "./src/contexts/auth0Context";
 
 export const wrapRootElement = ({ element }) => {
-  return <Auth0Provider>{element}</Auth0Provider>
-}
+  return <Auth0Provider>{element}</Auth0Provider>;
+};
