+++ frontend/lint-staged.config.js
@@ -1,7 +1,3 @@
 module.exports = {
-  "*.js": "prettier --write",
-  "*.ts?(x)": filenames =>
-    `concurrently "prettier --parser typescript --write ${filenames.join(
-      " "
-    )}" "npm run check-types"`,
+  "*.{js,jsx,ts,tsx,md,html,css}": "npx prettier --write",
 };
