+++ frontend/package.json
@@ -47,8 +47,9 @@
     "eslint-plugin-sonarjs": "^0.5.0",
     "gatsby-plugin-asset-path": "^3.0.2",
     "gatsby-plugin-root-import": "^2.0.5",
-    "lint-staged": "^10.2.11",
-    "prettier": "^1.19.1",
+    "husky": "^4.3.0",
+    "lint-staged": "^10.3.0",
+    "prettier": "2.1.1",
     "prettier-plugin-organize-imports": "^1.1.0",
     "stylelint": "^13.6.1",
     "stylelint-config-recommended": "^3.0.0",
@@ -68,7 +69,8 @@
     "clean": "gatsby clean",
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1",
     "type-check": "tsc --noEmit",
-    "lint": "concurrently \"node_modules/.bin/eslint src/**/*.{ts,tsx} --quiet --fix\" \"node_modules/.bin/stylelint --fix '**/*.{js,ts,tsx,css}'\"  \"npm run type-check\""
+    "lint": "concurrently \"node_modules/.bin/eslint src/**/*.{ts,tsx} --quiet --fix\" \"node_modules/.bin/stylelint --fix '**/*.{js,ts,tsx,css}'\"  \"npm run type-check\"",
+    "prettier-check": "npx prettier --check ."
   },
   "repository": {
     "type": "git",
