+++ frontend/src/common/localStorage/get.ts
@@ -1,5 +1,7 @@
+import { isSSR } from "../utils/isSSR";
+
 export function get(key: string): string | null {
-  if (typeof window === "undefined") return null;
+  if (isSSR()) return null;
 
   return window?.localStorage?.getItem(key) || null;
 }
