+++ frontend/src/common/utils/isSSR.ts
@@ -1,0 +1,3 @@
+export function isSSR(): boolean {
+  return typeof window === "undefined";
+}
