+++ frontend/src/components/Footer/index.tsx
@@ -4,9 +4,7 @@ import { MainWrapper, Wrapper } from "./style";
 const Footer: FC = () => {
   return (
     <Wrapper>
-        <MainWrapper>
-            © 2020 cellxgene
-        </MainWrapper>
+      <MainWrapper>© 2020 cellxgene</MainWrapper>
     </Wrapper>
   );
 };
