+++ frontend/src/components/Footer/style.ts
@@ -1,22 +1,19 @@
 import { fontStyle } from "src/components/common/theme";
-import { LAYOUT } from "src/components/common/layout";
 import styled from "styled-components";
+import { layout } from "../common/layout";
 
 export const Wrapper = styled.div`
   width: 100%;
-  margin-right: ${LAYOUT.NAME_MARGIN}%;
   height: 56px;
   padding-left: 15px;
   display: flex;
   align-items: center;
-  border-top: 1px solid #DBDCDD;
+  border-top: 1px solid #dbdcdd;
 `;
 
 export const MainWrapper = styled.div`
   ${fontStyle}
-  max-width: 1400px;
-  min-width: 800px;
-  width: 100%;
+  ${layout}
   padding: 0 15px;
   margin: 0 auto;
 `;
