+++ frontend/src/components/Layout/style.ts
@@ -1,4 +1,5 @@
 import styled from "styled-components";
+import { layout } from "../common/layout";
 
 export const Wrapper = styled.div`
   display: flex;
@@ -6,9 +7,7 @@ export const Wrapper = styled.div`
 `;
 
 export const MainWrapper = styled.div`
-  max-width: 1400px;
-  min-width: 800px;
-  width: 100%;
+  ${layout}
   min-height: 83vh;
   padding: 15px 25px;
   margin: 0 auto;
