+++ frontend/src/components/ProjectList/common/layout.ts
@@ -1,0 +1,8 @@
+// (thuang): Unit is %
+export enum LAYOUT {
+  NAME = 58.33,
+  NAME_MARGIN = 4.17,
+  VIEW = 14.81,
+  VIEW_MARGIN = 6.02,
+  INFO = 16.67,
+}
