+++ frontend/src/components/ProjectList/components/Dataset/components/MoreInformation/style.ts
@@ -1,4 +1,4 @@
-import { LAYOUT } from "src/components/common/layout";
+import { LAYOUT } from "src/components/ProjectList/common/layout";
 import styled from "styled-components";
 
 export const Wrapper = styled.div`
