+++ frontend/src/components/ProjectList/components/Dataset/components/ViewDataset/style.ts
@@ -1,5 +1,5 @@
 import { fontStyle } from "src/components/common/theme";
-import { LAYOUT } from "src/components/common/layout";
+import { LAYOUT } from "src/components/ProjectList/common/layout";
 import styled from "styled-components";
 
 export const Wrapper = styled.div`
@@ -18,9 +18,9 @@ export const SelectWrapper = styled.div`
 export const StyledAnchor = styled.a`
   ${fontStyle}
   text-decoration: none;
-  color: #0076DC;
+  color: #0076dc;
   background-color: white;
-  border: 1px solid #0076DC;
+  border: 1px solid #0076dc;
   box-shadow: inset 0px -1px 1px rgba(16, 22, 26, 0.2);
   border-radius: 3px;
   appearance: none;
