+++ frontend/src/components/ProjectList/components/Dataset/style.ts
@@ -1,5 +1,5 @@
 import { fontStyle } from "src/components/common/theme";
-import { LAYOUT } from "src/components/common/layout";
+import { LAYOUT } from "src/components/ProjectList/common/layout";
 import styled from "styled-components";
 
 export const Wrapper = styled.div`
