+++ frontend/src/components/common/layout.ts
@@ -1,8 +1,7 @@
-// (thuang): Unit is %
-export enum LAYOUT {
-    NAME = 58.33,
-    NAME_MARGIN = 4.17,
-    VIEW = 14.81,
-    VIEW_MARGIN = 6.02,
-    INFO = 16.67,
-}
+import { css } from "styled-components";
+
+export const layout = css`
+  max-width: 1400px;
+  min-width: 800px;
+  width: 100%;
+`;
